producer/requirements.txt
websocket-client==1.5.1
confluent-kafka==1.9.2